requirements.txt
beautifulsoup4==4.12.2
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
python-dotenv==1.0.0
requests==2.30.0
